+++ postgresql-core/src/main/java/com/github/starnowski/posmulten/postgresql/core/db/operations/DropOperationsProcessor.java
@@ -24,6 +24,7 @@
 package com.github.starnowski.posmulten.postgresql.core.db.operations;
 
 import com.github.starnowski.posmulten.postgresql.core.common.SQLDefinition;
+import com.github.starnowski.posmulten.postgresql.core.db.operations.exceptions.ValidationDatabaseOperationsException;
 
 import javax.sql.DataSource;
 import java.sql.Connection;
@@ -32,16 +33,40 @@ import java.sql.Statement;
 import java.util.LinkedList;
 import java.util.List;
 
+/**
+ * Database operation process that executes drop scripts.
+ */
 public class DropOperationsProcessor implements IDatabaseOperationsProcessor {
+    /**
+     * Executes drop scripts for dataSource object.
+     * Process attempts to establish a connection with the data source that this DataSource object represents.
+     * At the end of operation the established connection object is going to be closed.
+     * @param dataSource Datasource object
+     * @param sqlDefinitions list of sql definitions objects
+     * @throws SQLException
+     * @throws ValidationDatabaseOperationsException
+     */
     @Override
-    public void run(DataSource dataSource, List<SQLDefinition> sqlDefinitions) throws SQLException {
+    public void run(DataSource dataSource, List<SQLDefinition> sqlDefinitions) throws SQLException, ValidationDatabaseOperationsException {
         try (Connection connection = dataSource.getConnection()) {
-            LinkedList<SQLDefinition> stack = new LinkedList<>();
-            sqlDefinitions.forEach(stack::push);
-            for (SQLDefinition sqlDefinition : stack) {
-                Statement statement = connection.createStatement();
-                statement.execute(sqlDefinition.getDropScript());
-            }
+            this.run(connection, sqlDefinitions);
+        }
+    }
+
+    /**
+     * Executes drop scripts for passed connection object.
+     * @param connection Connection object
+     * @param sqlDefinitions list of sql definitions objects
+     * @throws SQLException
+     * @throws ValidationDatabaseOperationsException
+     */
+    @Override
+    public void run(Connection connection, List<SQLDefinition> sqlDefinitions) throws SQLException, ValidationDatabaseOperationsException {
+        LinkedList<SQLDefinition> stack = new LinkedList<>();
+        sqlDefinitions.forEach(stack::push);
+        for (SQLDefinition sqlDefinition : stack) {
+            Statement statement = connection.createStatement();
+            statement.execute(sqlDefinition.getDropScript());
         }
     }
 }