+++ CHANGELOG.md
@@ -9,6 +9,7 @@ https://keepachangelog.com/en/1.0.0/
 https://www.markdownguide.org/basic-syntax/
 
 * [Unreleased](#unreleased)
+* [0.7.1](#071---2023-07-02)
 * [0.7.0](#070---2023-06-28)
 * [0.6.0](#060---2023-01-14)
 * [0.5.6](#056---2022-11-04)
@@ -27,6 +28,18 @@ https://www.markdownguide.org/basic-syntax/
 * [0.2.0](#020---2020-11-14)
 
 ## [Unreleased]
+
+## [0.7.1] - 2023-07-02
+
+### Added
+
+- Added enhancement to DatabaseOperationExecutor and IDatabaseOperationsProcessor types [288](https://github.com/starnowski/posmulten/issues/288)
+    - Added method com.github.starnowski.posmulten.postgresql.core.db.DatabaseOperationExecutor#execute(Connection, List<SQLDefinition> sqlDefinitions, DatabaseOperationType). [288](https://github.com/starnowski/posmulten/issues/288)
+    - Added method com.github.starnowski.posmulten.postgresql.core.db.operations.CreateOperationsProcessor#run(Connection, List<SQLDefinition> sqlDefinitions). [288](https://github.com/starnowski/posmulten/issues/288)
+    - Added method com.github.starnowski.posmulten.postgresql.core.db.operations.DatabaseOperationsLoggerProcessor#run(Connection, List<SQLDefinition> sqlDefinitions). [288](https://github.com/starnowski/posmulten/issues/288)
+    - Added method com.github.starnowski.posmulten.postgresql.core.db.operations.DropOperationsProcessor#run(Connection, List<SQLDefinition> sqlDefinitions). [288](https://github.com/starnowski/posmulten/issues/288)
+    - Added method com.github.starnowski.posmulten.postgresql.core.db.operations.IDatabaseOperationsProcessor#run(Connection, List<SQLDefinition> sqlDefinitions). [288](https://github.com/starnowski/posmulten/issues/288)
+    - Added method com.github.starnowski.posmulten.postgresql.core.db.operations.ValidateOperationsProcessor#run(Connection, List<SQLDefinition> sqlDefinitions). [288](https://github.com/starnowski/posmulten/issues/288)
 
 ## [0.7.0] - 2023-06-28